created playground, style dummybox, have data folder
playground is used to be able to test what is needed to be done. Currently what is being tested is being able to show data from an array of objects instead of the static info we have.  style dummybox to give direction of how it could possibly look. Maybe look into shadows. have data folder for the data we have now. will probably be moved when you get the database