update not working
Having edit able to find the data from the edit modal that is put in by the user. Also have a state to hold that data as well as tack on the id and people that are connected with it. From there when I try to replace the data by the submit button, it breaks completely