Clean up (know the bug)
when you add a new event you aren't adding a people array of objects so when you try to push to it. It breaks cuz it does not exist. Really need to get people edit done so this issue can be fixed as well as add a people array that holds nothing when you add any event