3ish items left for full functionality
All notes are in the code. Whats left is really just:

 deleting people  (the frame work for what it is now possibly might not work)  

if seats === 0 then send a success 200 not a success 204 / if a person is deleted then seats go back up

other than that, that should be it

random note: possibly tying the index to the checkbox instead of an actual part of the object might fix the checkboxes. 

This will probably be my last update unless I somehow have freetime on monday before classes. I will try to help finish getting it connected after that y'all got it from there to find and fix bugs while skons gets the path for getting it on partybus.

Y'all got this team!

-LT