set new item up almost done
Still need to do 2 more things till add event is done the 2 things are weekly repeat and window alert. I commented out what still needs to be done on edit and delete above the API part of the that events code. (delete to do items above the delete function in the api, the to do items for edit events is above the update function in the api code area).